# docs/swagger/swagger.json
{
    "swagger":  "2.0",
    "info":  {
                 "contact":  {

                             },
                 "title":  "Order Processing API",
                 "description":  "Combined API documentation for auth, product, and order services."
             },
    "paths":  {
                  "/auth/login":  {
                                      "post":  {
                                                   "consumes":  [
                                                                    "application/json"
                                                                ],
                                                   "produces":  [
                                                                    "application/json"
                                                                ],
                                                   "tags":  [
                                                                "auth"
                                                            ],
                                                   "summary":  "Login user",
                                                   "parameters":  [
                                                                      {
                                                                          "description":  "Login payload",
                                                                          "name":  "payload",
                                                                          "in":  "body",
                                                                          "required":  true,
                                                                          "schema":  {
                                                                                         "$ref":  "#/definitions/main.AuthRequest"
                                                                                     }
                                                                      }
                                                                  ],
                                                   "responses":  {
                                                                     "200":  {
                                                                                 "description":  "OK",
                                                                                 "schema":  {
                                                                                                "$ref":  "#/definitions/main.authTokenDataResponse"
                                                                                            }
                                                                             },
                                                                     "400":  {
                                                                                 "description":  "Bad Request",
                                                                                 "schema":  {
                                                                                                "$ref":  "#/definitions/main.authEmptyResponse"
                                                                                            }
                                                                             },
                                                                     "401":  {
                                                                                 "description":  "Unauthorized",
                                                                                 "schema":  {
                                                                                                "$ref":  "#/definitions/main.authEmptyResponse"
                                                                                            }
                                                                             }
                                                                 }
                                               }
                                  },
                  "/auth/logout":  {
                                       "post":  {
                                                    "security":  [
                                                                     {
                                                                         "BearerAuth":  [

                                                                                        ]
                                                                     }
                                                                 ],
                                                    "produces":  [
                                                                     "application/json"
                                                                 ],
                                                    "tags":  [
                                                                 "auth"
                                                             ],
                                                    "summary":  "Logout user",
                                                    "responses":  {
                                                                      "200":  {
                                                                                  "description":  "OK",
                                                                                  "schema":  {
                                                                                                 "$ref":  "#/definitions/main.authEmptyResponse"
                                                                                             }
                                                                              },
                                                                      "401":  {
                                                                                  "description":  "Unauthorized",
                                                                                  "schema":  {
                                                                                                 "$ref":  "#/definitions/main.authEmptyResponse"
                                                                                             }
                                                                              }
                                                                  }
                                                }
                                   },
                  "/auth/refresh":  {
                                        "post":  {
                                                     "consumes":  [
                                                                      "application/json"
                                                                  ],
                                                     "produces":  [
                                                                      "application/json"
                                                                  ],
                                                     "tags":  [
                                                                  "auth"
                                                              ],
                                                     "summary":  "Refresh token pair",
                                                     "parameters":  [
                                                                        {
                                                                            "description":  "Refresh payload",
                                                                            "name":  "payload",
                                                                            "in":  "body",
                                                                            "required":  true,
                                                                            "schema":  {
                                                                                           "$ref":  "#/definitions/main.RefreshTokenRequest"
                                                                                       }
                                                                        }
                                                                    ],
                                                     "responses":  {
                                                                       "200":  {
                                                                                   "description":  "OK",
                                                                                   "schema":  {
                                                                                                  "$ref":  "#/definitions/main.authTokenDataResponse"
                                                                                              }
                                                                               },
                                                                       "400":  {
                                                                                   "description":  "Bad Request",
                                                                                   "schema":  {
                                                                                                  "$ref":  "#/definitions/main.authEmptyResponse"
                                                                                              }
                                                                               },
                                                                       "401":  {
                                                                                   "description":  "Unauthorized",
                                                                                   "schema":  {
                                                                                                  "$ref":  "#/definitions/main.authEmptyResponse"
                                                                                              }
                                                                               }
                                                                   }
                                                 }
                                    },
                  "/auth/register":  {
                                         "post":  {
                                                      "consumes":  [
                                                                       "application/json"
                                                                   ],
                                                      "produces":  [
                                                                       "application/json"
                                                                   ],
                                                      "tags":  [
                                                                   "auth"
                                                               ],
                                                      "summary":  "Register user",
                                                      "parameters":  [
                                                                         {
                                                                             "description":  "Register payload",
                                                                             "name":  "payload",
                                                                             "in":  "body",
                                                                             "required":  true,
                                                                             "schema":  {
                                                                                            "$ref":  "#/definitions/main.AuthRequest"
                                                                                        }
                                                                         }
                                                                     ],
                                                      "responses":  {
                                                                        "201":  {
                                                                                    "description":  "Created",
                                                                                    "schema":  {
                                                                                                   "$ref":  "#/definitions/main.authUserDataResponse"
                                                                                               }
                                                                                },
                                                                        "400":  {
                                                                                    "description":  "Bad Request",
                                                                                    "schema":  {
                                                                                                   "$ref":  "#/definitions/main.authEmptyResponse"
                                                                                               }
                                                                                },
                                                                        "500":  {
                                                                                    "description":  "Internal Server Error",
                                                                                    "schema":  {
                                                                                                   "$ref":  "#/definitions/main.authEmptyResponse"
                                                                                               }
                                                                                }
                                                                    }
                                                  }
                                     },
                  "/auth/validate":  {
                                         "get":  {
                                                     "security":  [
                                                                      {
                                                                          "BearerAuth":  [

                                                                                         ]
                                                                      }
                                                                  ],
                                                     "produces":  [
                                                                      "application/json"
                                                                  ],
                                                     "tags":  [
                                                                  "auth"
                                                              ],
                                                     "summary":  "Validate bearer token",
                                                     "responses":  {
                                                                       "200":  {
                                                                                   "description":  "OK",
                                                                                   "schema":  {
                                                                                                  "$ref":  "#/definitions/main.authValidationDataResponse"
                                                                                              }
                                                                               },
                                                                       "401":  {
                                                                                   "description":  "Unauthorized",
                                                                                   "schema":  {
                                                                                                  "$ref":  "#/definitions/main.authEmptyResponse"
                                                                                              }
                                                                               }
                                                                   }
                                                 }
                                     },
...
              },
    "definitions":  {
                        "main.AuthRequest":  {
                                                 "type":  "object",
                                                 "properties":  {
                                                                    "email":  {
                                                                                  "type":  "string"
                                                                              },
                                                                    "password":  {
                                                                                     "type":  "string"
                                                                                 }
                                                                }
                                             },
                        "main.AuthResponse":  {
                                                  "type":  "object",
                                                  "properties":  {
                                                                     "access_token":  {
                                                                                          "type":  "string"
                                                                                      },
                                                                     "refresh_token":  {
                                                                                           "type":  "string"
                                                                                       }
                                                                 }
                                              },
                        "main.AuthValidationResponse":  {
                                                            "type":  "object",
                                                            "properties":  {
                                                                               "role":  {
                                                                                            "type":  "string"
                                                                                        },
                                                                               "user_id":  {
                                                                                               "type":  "string"
                                                                                           }
                                                                           }
                                                        },
                        "main.RefreshTokenRequest":  {
                                                         "type":  "object",
                                                         "properties":  {
                                                                            "refresh_token":  {
                                                                                                  "type":  "string"
                                                                                              }
                                                                        }
                                                     },
                        "main.authEmptyResponse":  {
                                                       "type":  "object",
                                                       "properties":  {
                                                                          "message":  {
                                                                                          "type":  "string"
                                                                                      },
                                                                          "status":  {
                                                                                         "type":  "integer"
                                                                                     }
                                                                      }
                                                   },
                        "main.authTokenDataResponse":  {
                                                           "type":  "object",
                                                           "properties":  {
                                                                              "data":  {
                                                                                           "$ref":  "#/definitions/main.AuthResponse"
                                                                                       },
                                                                              "message":  {
                                                                                              "type":  "string"
                                                                                          },
                                                                              "status":  {
                                                                                             "type":  "integer"
                                                                                         }
                                                                          }
                                                       },
                        "main.authUserDataResponse":  {
                                                          "type":  "object",
                                                          "properties":  {
                                                                             "data":  {
                                                                                          "$ref":  "#/definitions/models.User"
                                                                                      },
                                                                             "message":  {
                                                                                             "type":  "string"
                                                                                         },
                                                                             "status":  {
                                                                                            "type":  "integer"
                                                                                        }
                                                                         }
                                                      },
                        "main.authValidationDataResponse":  {
                                                                "type":  "object",
                                                                "properties":  {
                                                                                   "data":  {
                                                                                                "$ref":  "#/definitions/main.AuthValidationResponse"
                                                                                            },
                                                                                   "message":  {
                                                                                                   "type":  "string"
                                                                                               },
                                                                                   "status":  {
                                                                                                  "type":  "integer"
                                                                                              }
                                                                               }
                                                            },
                        "models.User":  {
                                            "type":  "object",
                                            "properties":  {
                                                               "email":  {
                                                                             "type":  "string"
                                                                         },
                                                               "id":  {
                                                                          "type":  "string"
                                                                      },
                                                               "role":  {
                                                                            "type":  "string"
                                                                        }
                                                           }
                                        },
...
                    },
    "securityDefinitions":  {

                            },
...
}